kiwi: Remove low_latency permission, we aren't low latency
Using Superpowered Latency Test, on stock I get about 80ms.  On
Lineage 14.1, with the low_latency permission and our reduced
period size, I get around 55ms but it often fails and distorts
the sound.  Removing the period and the permission indicates that
it is really around 65ms which is slower than the required 45ms
latency to be truly low_latency.

Change-Id: I9c82a2bb271b5dccfddabad064b67d26a1f53b0c